Fix export problem
