+++ lib/src/EventSystem/EventSystem.ts
@@ -1,75 +1,10 @@
-import { mat4, vec2 } from "gl-matrix"
+import { vec2 } from "gl-matrix"
 import { IRect } from "../helpers/geometry"
-
-export interface HitAreaEventHandler<T = unknown> {
-  (event: HitAreaEvent<T>): void
-}
+import { HitArea } from "./HitArea"
+import { HitAreaEvent } from "./HitAreaEvent"
 
 // Union type for mouse and pointer events
 export type InputEvent = MouseEvent | PointerEvent
-
-// Function to extract coordinates from either MouseEvent or PointerEvent
-function getEventCoordinates(event: InputEvent): {
-  clientX: number
-  clientY: number
-} {
-  return {
-    clientX: event.clientX,
-    clientY: event.clientY,
-  }
-}
-
-export class HitAreaEvent<T = unknown> {
-  public readonly nativeEvent: InputEvent
-  public readonly point: vec2
-  public readonly data?: T
-
-  private propagationStopped = false
-  private defaultPrevented = false
-
-  constructor(nativeEvent: InputEvent, point: vec2, data?: T) {
-    this.nativeEvent = nativeEvent
-    this.point = point
-    this.data = data
-  }
-
-  stopPropagation(): void {
-    this.propagationStopped = true
-  }
-
-  preventDefault(): void {
-    this.defaultPrevented = true
-    this.nativeEvent.preventDefault()
-  }
-
-  get isPropagationStopped(): boolean {
-    return this.propagationStopped
-  }
-
-  get isDefaultPrevented(): boolean {
-    return this.defaultPrevented
-  }
-}
-
-export interface HitArea<T = unknown> {
-  id: string
-  bounds: IRect
-  transform: mat4
-  zIndex: number
-  onMouseDown?: HitAreaEventHandler<T>
-  onMouseUp?: HitAreaEventHandler<T>
-  onMouseMove?: HitAreaEventHandler<T>
-  onMouseEnter?: HitAreaEventHandler<T>
-  onMouseLeave?: HitAreaEventHandler<T>
-  onClick?: HitAreaEventHandler<T>
-  onPointerDown?: HitAreaEventHandler<T>
-  onPointerUp?: HitAreaEventHandler<T>
-  onPointerMove?: HitAreaEventHandler<T>
-  onPointerEnter?: HitAreaEventHandler<T>
-  onPointerLeave?: HitAreaEventHandler<T>
-  onPointerCancel?: HitAreaEventHandler<T>
-  data?: T
-}
 
 export interface GLCanvasEventHandler {
   (event: InputEvent): void
@@ -84,41 +19,6 @@ export interface GLCanvasEventHandlers {
   onPointerUp?: GLCanvasEventHandler
   onPointerMove?: GLCanvasEventHandler
   onPointerCancel?: GLCanvasEventHandler
-}
-
-// Pure functions (file-private)
-function getCanvasPoint(event: InputEvent, canvas: HTMLCanvasElement): vec2 {
-  const rect = canvas.getBoundingClientRect()
-  const coords = getEventCoordinates(event)
-  const x = coords.clientX - rect.left
-  const y = coords.clientY - rect.top
-  return vec2.fromValues(x, y)
-}
-
-function isPointInBounds(point: vec2, bounds: IRect): boolean {
-  return (
-    point[0] >= bounds.x &&
-    point[0] <= bounds.x + bounds.width &&
-    point[1] >= bounds.y &&
-    point[1] <= bounds.y + bounds.height
-  )
-}
-
-function findTopHitArea(
-  hitAreas: HitArea<any>[],
-  point: vec2,
-): HitArea<any> | null {
-  const sortedHitAreas = [...hitAreas].sort(
-    (a, b) => (b.zIndex || 0) - (a.zIndex || 0),
-  )
-
-  for (const hitArea of sortedHitAreas) {
-    if (isPointInBounds(point, hitArea.bounds)) {
-      return hitArea
-    }
-  }
-
-  return null
 }
 
 export class EventSystem {
@@ -192,7 +92,7 @@ export class EventSystem {
     >,
     canvasEventType: keyof GLCanvasEventHandlers,
   ): boolean {
-    const point = getCanvasPoint(event, canvas)
+    const point = getLocalPoint(event, canvas)
     const hitArea = findTopHitArea(this.hitAreas, point)
 
     if (hitArea) {
@@ -219,7 +119,7 @@ export class EventSystem {
     leaveEventType: keyof Pick<HitArea<any>, "onMouseLeave" | "onPointerLeave">,
     canvasEventType: keyof GLCanvasEventHandlers,
   ): boolean {
-    const point = getCanvasPoint(event, canvas)
+    const point = getLocalPoint(event, canvas)
     const hitArea = findTopHitArea(this.hitAreas, point)
 
     if (hitArea !== this.hoveredHitArea) {
@@ -311,4 +211,49 @@ export class EventSystem {
       "onPointerCancel",
     )
   }
+}
+
+// Function to extract coordinates from either MouseEvent or PointerEvent
+function getEventCoordinates(event: InputEvent): {
+  clientX: number
+  clientY: number
+} {
+  return {
+    clientX: event.clientX,
+    clientY: event.clientY,
+  }
+}
+
+function getLocalPoint(event: InputEvent, element: HTMLElement): vec2 {
+  const rect = element.getBoundingClientRect()
+  const coords = getEventCoordinates(event)
+  const x = coords.clientX - rect.left
+  const y = coords.clientY - rect.top
+  return vec2.fromValues(x, y)
+}
+
+function isPointInBounds(point: vec2, bounds: IRect): boolean {
+  return (
+    point[0] >= bounds.x &&
+    point[0] <= bounds.x + bounds.width &&
+    point[1] >= bounds.y &&
+    point[1] <= bounds.y + bounds.height
+  )
+}
+
+function findTopHitArea(
+  hitAreas: HitArea<any>[],
+  point: vec2,
+): HitArea<any> | null {
+  const sortedHitAreas = [...hitAreas].sort(
+    (a, b) => (b.zIndex || 0) - (a.zIndex || 0),
+  )
+
+  for (const hitArea of sortedHitAreas) {
+    if (isPointInBounds(point, hitArea.bounds)) {
+      return hitArea
+    }
+  }
+
+  return null
 }
