+++ lib/src/index.ts
@@ -1,3 +1,5 @@
+export * from "./EventSystem/EventSystem"
+export * from "./EventSystem/HitAreaEvent"
 export * from "./GLCanvas/GLCanvas"
 export * from "./GLNode/BorderedCircles/BorderedCircles"
 export * from "./GLNode/BorderedRectangles/BorderedRectangles"
@@ -7,12 +9,11 @@ export * from "./GLNode/GLNodeInternal"
 export * from "./GLNode/HitArea"
 export * from "./GLNode/Rectangles/Rectangles"
 export * from "./GLNode/Transform"
-export * from "./EventSystem/EventSystem"
 export * from "./helpers/geometry"
 export * from "./helpers/polygon"
 export * from "./hooks/useEventSystem"
-export * from "./hooks/useTransform"
 export * from "./hooks/useRenderer"
+export * from "./hooks/useTransform"
 export * from "./Renderer/Renderer"
 export * from "./Shader/Shader"
 export * from "./Shader/Uniform"
