+++ lib/tsconfig.json
@@ -21,5 +21,6 @@
     "skipLibCheck": true,
     "moduleResolution": "node"
   },
-  "include": ["./src"]
+  "include": ["./src"],
+  "exclude": ["node_modules", "dist", "**/*.test.ts", "**/*.test.tsx"]
 }
